requirements.txt
flask
flask-cors
langchain-huggingface
langchain-core
python-dotenv